Nombres de clases ajustados
