+++ Negocio/src/main/java/com/ciberdix/th/controllers/refactor/ListasEstadosCivilesRefactorController.java
@@ -1,10 +1,9 @@
 package com.ciberdix.th.controllers.refactor;
 
 import com.ciberdix.th.config.Globales;
-import com.ciberdix.th.model.refactor.ListasGeneros;
+import com.ciberdix.th.model.refactor.ListasEstadosCiviles;
 import java.util.Arrays;
 import java.util.List;
-import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
@@ -15,19 +14,19 @@ import org.springframework.web.client.RestTemplate;
  *
  * @author Roberto Chajin Ortiz
  */
-@RestController
 @CrossOrigin
-@Transactional
-@RequestMapping("/api/generos")
-public class GenerosControlador {
+@RestController
+@RequestMapping("/api/estadosCiviles")
+public class ListasEstadosCivilesRefactorController {
+
     Globales globales = new Globales();
     private String serviceUrl = globales.getUrl();
     
     @RequestMapping(method = RequestMethod.GET)
-    List<ListasGeneros> listarGeneros() {
+    List<ListasEstadosCiviles> listarEstadosCiviles() {
         RestTemplate restTemplate = new RestTemplate();
-        ListasGeneros[] generos = restTemplate.getForObject(serviceUrl + "api/generos", ListasGeneros[].class);
+        ListasEstadosCiviles[] estados = restTemplate.getForObject(serviceUrl + "api/estadosCiviles", ListasEstadosCiviles[].class);
 
-        return Arrays.asList(generos);
-    }       
+        return Arrays.asList(estados);
+    }    
 }
