+++ Negocio/src/main/java/com/ciberdix/th/controllers/refactor/ListasTiposPersonasRefactorController.java
@@ -1,7 +1,7 @@
 package com.ciberdix.th.controllers.refactor;
 
 import com.ciberdix.th.config.Globales;
-import com.ciberdix.th.model.refactor.Ocupaciones;
+import com.ciberdix.th.model.refactor.ListasTiposPersonas;
 import java.util.Arrays;
 import java.util.List;
 import org.springframework.web.bind.annotation.CrossOrigin;
@@ -16,17 +16,17 @@ import org.springframework.web.client.RestTemplate;
  */
 @CrossOrigin
 @RestController
-@RequestMapping("/api/ocupaciones")
-public class OcupacionesControlador {
+@RequestMapping("/api/tiposPersonas")
+public class ListasTiposPersonasRefactorController {
 
     Globales globales = new Globales();
     private String serviceUrl = globales.getUrl();
-    
-    @RequestMapping(method = RequestMethod.GET)
-    List<Ocupaciones> listarOcupaciones() {
-        RestTemplate restTemplate = new RestTemplate();
-        Ocupaciones[] ocupaciones = restTemplate.getForObject(serviceUrl + "api/ocupaciones", Ocupaciones[].class);
 
-        return Arrays.asList(ocupaciones);
+    @RequestMapping(method = RequestMethod.GET)
+    List<ListasTiposPersonas> listarTiposPersonas() {
+        RestTemplate restTemplate = new RestTemplate();
+        ListasTiposPersonas[] tiposPersonas = restTemplate.getForObject(serviceUrl + "api/tiposPersonas", ListasTiposPersonas[].class);
+
+        return Arrays.asList(tiposPersonas);
     }    
 }
