+++ Negocio/src/main/java/com/ciberdix/th/controllers/refactor/OcupacionesRefactorController.java
@@ -1,7 +1,7 @@
 package com.ciberdix.th.controllers.refactor;
 
 import com.ciberdix.th.config.Globales;
-import com.ciberdix.th.model.refactor.ListasFactoresRh;
+import com.ciberdix.th.model.refactor.Ocupaciones;
 import java.util.Arrays;
 import java.util.List;
 import org.springframework.web.bind.annotation.CrossOrigin;
@@ -16,17 +16,17 @@ import org.springframework.web.client.RestTemplate;
  */
 @CrossOrigin
 @RestController
-@RequestMapping("/api/factorRh")
-public class FactorRhControlador {
+@RequestMapping("/api/ocupaciones")
+public class OcupacionesRefactorController {
 
     Globales globales = new Globales();
     private String serviceUrl = globales.getUrl();
     
     @RequestMapping(method = RequestMethod.GET)
-    List<ListasFactoresRh> listarFactorRh() {
+    List<Ocupaciones> listarOcupaciones() {
         RestTemplate restTemplate = new RestTemplate();
-        ListasFactoresRh[] factores = restTemplate.getForObject(serviceUrl + "api/factorRh", ListasFactoresRh[].class);
+        Ocupaciones[] ocupaciones = restTemplate.getForObject(serviceUrl + "api/ocupaciones", Ocupaciones[].class);
 
-        return Arrays.asList(factores);
-    }     
+        return Arrays.asList(ocupaciones);
+    }    
 }
